m03/requirements.txt
# Requirements for Module 03 - QA Bot with Local Models
# Optimized for NVIDIA 4090 GPU

# Core LangChain dependencies
langchain==0.2.11
langchain-core>=0.2.23,<0.3.0
langchain-community==0.2.10
langchain-text-splitters==0.2.2

# Local LLM and Embedding alternatives (GPU-optimized)
transformers>=4.40.0
torch>=2.2.0  # Will auto-detect CUDA
torchvision>=0.17.0
torchaudio>=2.2.0
accelerate>=0.30.0
bitsandbytes>=0.43.0  # For quantization
sentence-transformers>=3.0.1

# Vector database dependencies
chromadb==0.4.24
faiss-cpu==1.8.0  # CPU version for compatibility

# Document loading and processing
pypdf==4.3.1
PyMuPDF==1.24.5  # Alternative PDF processor

# Web interface
gradio>=4.44.0

# Additional utilities
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
loguru>=0.7.0

# PNG generation for submission
Pillow>=10.0.0
pygments>=2.15.0

# Optional: Ollama support (for local LLMs)
# ollama  # Install separately if using Ollama
